requirements.txt
pytest==8.0.2
nose2==0.14.1
pylint==3.1.0
requests==2.31.0
python-dotenv==1.0.1
behave==1.2.6